requirements.txt
# Core dependencies - optimized versions
streamlit>=1.32.0
python-dotenv>=1.0.0
chromadb>=0.4.22
langchain>=0.1.10
langchain-community>=0.0.25
langchain-openai>=0.0.8
sentence-transformers>=2.6.1

# Document processing - optimized
PyPDF2>=3.0.1
pdf2image>=1.17.0
pillow>=10.2.0
python-docx>=1.1.0

# OCR and computer vision - performance optimized
pytesseract>=0.3.10
opencv-python>=4.9.0.80
easyocr>=1.7.1

# Table and chart processing - efficient versions
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.3
seaborn>=0.13.2
plotly>=5.18.0
tabula-py>=2.9.0
# Note: camelot-py requires additional system dependencies (cv2, etc.)
# camelot-py[cv]>=0.11.0

# Layout analysis (optional - will use basic CV methods if not available)
# layoutparser[paddlepaddle]>=0.3.4
# detectron2>=0.6

# API and utilities - optimized
openai>=1.12.0
google-generativeai>=0.4.0
requests>=2.31.0
python-multipart>=0.0.7
aiofiles>=23.2.1

# Performance and optimization
asyncio-throttle>=1.0.2
memory-profiler>=0.61.0
psutil>=5.9.8
cachetools>=5.3.2

# Development and testing
pytest>=8.0.0
black>=24.1.0
flake8>=7.0.0
